Extract benchmark report chart rendering
